+++ lib/app/timeline/settings/edit/edit_timeline_settings_dialog.dart
@@ -1,3 +1,4 @@
+import 'package:easy_dispose_provider/easy_dispose_provider.dart';
 import 'package:fedi/app/auth/instance/current/current_auth_instance_bloc.dart';
 import 'package:fedi/app/instance/location/instance_location_model.dart';
 import 'package:fedi/app/settings/settings_dialog.dart';
@@ -12,7 +13,6 @@ import 'package:fedi/app/timeline/timeline_label_extension.dart';
 import 'package:fedi/app/timeline/timeline_model.dart';
 import 'package:fedi/app/ui/async/fedi_async_init_loading_widget.dart';
 import 'package:fedi/app/web_sockets/settings/web_sockets_settings_bloc.dart';
-import 'package:easy_dispose_provider/easy_dispose_provider.dart';
 import 'package:fedi/generated/l10n.dart';
 import 'package:fedi/local_preferences/local_preferences_service.dart';
 import 'package:fedi/pleroma/api/instance/pleroma_api_instance_model.dart';
@@ -100,7 +100,10 @@ TimelineLocalPreferenceBloc _createTimelinePreferencesBloc(
   BuildContext context,
   Timeline timeline,
 ) {
-  var localPreferencesService = ILocalPreferencesService.of(context);
+  var localPreferencesService = ILocalPreferencesService.of(
+    context,
+    listen: false,
+  );
   var currentAuthInstanceBloc =
       ICurrentAuthInstanceBloc.of(context, listen: false);
   var currentInstance = currentAuthInstanceBloc.currentInstance!;