+++ CHANGELOG.md
@@ -17,6 +17,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 * guest mode fixes for media
 * fix push notifications on mastodon
 * fix async_dialog.dart: don't crash when performing async operation without actual dialog
+* edit_timeline_settings_dialog.dart: don't listen provider in button callback
   
 
 ## [2.6.0] [internal] 2021 July 10